refactor out parse_data into separate file
